requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Snowflake connector
snowflake-connector-python>=3.7.0

# Data processing - Updated for Python 3.13 compatibility
pandas>=2.2.0
numpy>=1.26.0

# JSON handling
orjson>=3.9.10

# HTTP requests
requests>=2.31.0

# Cryptography for Snowflake JWT - Updated for Python 3.13
cryptography>=41.0.0

# Date/time handling
python-dateutil>=2.8.2

# Logging
structlog>=23.2.0

# Redis (for API caching)
redis>=5.0.1

# Environment variable handling
python-dotenv>=1.0.0

# Additional dependencies for testing
pytest>=7.0.0
pytest-cov>=4.0.0
